pyproject.toml
[tool.poetry]
name = "create-poll-pages"
version = "0.1.0"
description = ""
authors = ["Yuki Watanabe <ukwhatn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
wikidot = "3.1.0.dev5"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
